Refatoração Controller, criação de Objeto Response e add novos campos na entity

+++ src/main/java/br/com/tarz/pelandoapi/repository/PromocaoRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 import br.com.tarz.pelandoapi.entity.Promocao;
 
-public interface PromocaoRepository extends JpaRepository<Promocao, Integer>, JpaSpecificationExecutor<Promocao>{
+public interface PromocaoRepository extends JpaRepository<Promocao, Long>, JpaSpecificationExecutor<Promocao>{
 
 }
